refactoring + bug fixes
