+++ worker/service/worker/migration_workers/__init__.py
@@ -1,0 +1,3 @@
+from .reindex import reindex
+from .v071.validation_schema_reindex import validation_schema_reindex as validation_schema_reindex_071
+from .v071.user_reindex import user_reindex as user_reindex_071
