+++ worker/service/worker/migration_workers/utils/migration_error.py
@@ -1,0 +1,2 @@
+class MigrationError(Exception):
+    pass
